refactor: 안쓰는 변수 제거
