+++ src/main/java/lotto/generatepolicy/GeneratePolicy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.generatepolicy;
 
 public interface GeneratePolicy {
