+++ src/test/java/lotto/domain/LottoTest.java
@@ -5,6 +5,7 @@ import static lotto.domain.Lotto.makeLotto;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.generatepolicy.RandomGeneratePolicy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
